Fix ConcurrentModification on startup
