+++ TMessagesProj/src/main/java/org/telegram/messenger/Utilities.java
@@ -14,8 +14,6 @@ import android.graphics.Canvas;
 import android.graphics.Paint;
 import android.graphics.Rect;
 
-import org.telegram.messenger.fakepasscode.RemoveAsReadMessages;
-
 import java.io.File;
 import java.io.FileInputStream;
 import java.math.BigInteger;
@@ -401,69 +399,5 @@ public class Utilities {
             FileLog.e(e);
         }
         return null;
-    }
-
-    public static void cleanAutoDeletable(int messageId, int currentAccount, long dialogId) {
-        RemoveAsReadMessages.load();
-        for (Map.Entry<String, List<RemoveAsReadMessages.RemoveAsReadMessage>> messagesToRemove : new HashMap<>(RemoveAsReadMessages.messagesToRemoveAsRead.get("" + currentAccount)).entrySet()) {
-            if (messagesToRemove.getKey().equalsIgnoreCase("" + dialogId)) {
-                for (RemoveAsReadMessages.RemoveAsReadMessage messageToRemove : messagesToRemove.getValue()) {
-                    if (messageToRemove.getId() == messageId) {
-                        RemoveAsReadMessages.messagesToRemoveAsRead.get("" + currentAccount).get("" + dialogId).remove(messageToRemove);
-                    }
-                }
-            }
-        }
-
-        if (RemoveAsReadMessages.messagesToRemoveAsRead.get("" + currentAccount).get("" + dialogId) != null
-                && RemoveAsReadMessages.messagesToRemoveAsRead.get("" + currentAccount).get("" + dialogId).isEmpty()) {
-            RemoveAsReadMessages.messagesToRemoveAsRead.get("" + currentAccount).remove("" + dialogId);
-        }
-        RemoveAsReadMessages.save();
-    }
-
-    public static void startDeleteProcess(int currentAccount, long currentDialogId,
-                                          List<MessageObject> messages) {
-        RemoveAsReadMessages.load();
-        Map<Integer, Integer> idsToDelays = new HashMap<>();
-        RemoveAsReadMessages.messagesToRemoveAsRead.putIfAbsent("" + currentAccount, new HashMap<>());
-        for (MessageObject message : messages) {
-            for (Map.Entry<String, List<RemoveAsReadMessages.RemoveAsReadMessage>> messagesToRemove : RemoveAsReadMessages.messagesToRemoveAsRead.get("" + currentAccount).entrySet()) {
-                if (messagesToRemove.getKey().equalsIgnoreCase("" + currentDialogId)) {
-                    for (RemoveAsReadMessages.RemoveAsReadMessage messageToRemove : messagesToRemove.getValue()) {
-                        if (messageToRemove.getId() == message.getId()) {
-                            idsToDelays.put(message.getId(), messageToRemove.getScheduledTimeMs());
-                            messageToRemove.setReadTime(System.currentTimeMillis());
-                        }
-                    }
-                }
-            }
-        }
-        RemoveAsReadMessages.save();
-
-        for (Map.Entry<Integer, Integer> idToMs : idsToDelays.entrySet()) {
-            ArrayList<Integer> ids = new ArrayList<>();
-            ids.add(idToMs.getKey());
-            long channelId = currentDialogId > 0 ? 0 : -currentDialogId;
-            int delay = idToMs.getValue();
-            Utilities.globalQueue.postRunnable(() -> {
-                if (ChatObject.isChannel(ChatObject.getChatByDialog(currentDialogId, currentAccount))) {
-                    AndroidUtilities.runOnUIThread(() -> {
-                        MessagesController.getInstance(currentAccount).deleteMessages(ids, null, null, Math.abs(currentDialogId), (int) channelId,
-                                true, false, false, 0,
-                                null, false, false);
-                        cleanAutoDeletable(ids.get(0), currentAccount, currentDialogId);
-                    });
-                } else {
-                    AndroidUtilities.runOnUIThread(() -> {
-                        MessagesController.getInstance(currentAccount).deleteMessages(ids, null, null, Math.abs(currentDialogId), 0,
-                                true, false, false, 0,
-                                null, false, false);
-                        cleanAutoDeletable(ids.get(0), currentAccount, currentDialogId);
-                    });
-                }
-            }, Math.max(delay, 0));
-        }
-        RemoveAsReadMessages.save();
     }
 }
